refactor : 타이머 부모 컴포넌트로 위치
