+++ src/components/GameCard.tsx
@@ -1,44 +1,20 @@
 import { motion } from "framer-motion";
-import { useState, useEffect } from "react";
 import { useAppDispatch } from "../hooks/useAppDispatch";
-import { setAllFliedTrue, setFlipCard } from "../slice/pokemonSlice";
+import { setFlipCard } from "../slice/pokemonSlice";
 
 const GameCard: React.FC<{
   isFlied: boolean;
   pokemonName: string;
   number: number;
   uniqueId: string;
-}> = ({ isFlied, pokemonName, number, uniqueId }) => {
-  const [isClickEnabled, setIsClickEnabled] = useState(false); // 클릭 활성화 상태
-  const [selectCards, setSelectCards] = useState<
-    { pokemonName: string; uniqueId: string }[]
-  >([]);
-
+  isClickEnabled: boolean;
+}> = ({ isFlied, pokemonName, number, uniqueId, isClickEnabled }) => {
   const dispatch = useAppDispatch();
-
-  useEffect(() => {
-    const flipTimer = setTimeout(() => {
-      dispatch(setAllFliedTrue()); // 3초 후 카드가 setFlipped뒤집힘
-    }, 3000);
-
-    const clickTimer = setTimeout(() => {
-      setIsClickEnabled(true); // 4초 후 클릭 가능
-    }, 4000);
-
-    return () => {
-      clearTimeout(flipTimer);
-      clearTimeout(clickTimer); // 타이머 정리
-    };
-  }, [dispatch]);
-
   const handleCardClick = () => {
-    console.log(isFlied);
-    // Only allow adding the card to the selectCards if clicking is enabled
     if (isClickEnabled) {
       dispatch(setFlipCard(uniqueId));
     }
   };
-  console.log(selectCards);
   return (
     <>
       <motion.div
